requirements.txt
fastapi~=0.115.12
supabase~=2.15.0
python-dotenv~=1.1.0
ollama~=0.4.8
pillow~=11.2.1
openai~=1.76.0